Adds Support for Generating Passwords
Just printing passwords in the logs right now since it's difficult to test
due to it's randomness. Passwords for SQL Servers are not yet set, nor are
they stored in Key Vault.

Added the libraries for Key Vault, wrote code to get the vault uri from
settings, but then comment it out until we can confirm the new passwords
are legitimate.